Share encrypted replica runtime with mobile
