+++ crates/db-sync/src/lib.rs
@@ -1,0 +1,11 @@
+#![forbid(unsafe_code)]
+
+mod e2ee_sync;
+mod replica_sync;
+mod witness;
+mod witness_watch;
+
+pub use e2ee_sync::{E2eeSyncHook, ReplicaSyncOutcome, ReplicaSyncStatus};
+pub use replica_sync::{ReplicaSyncTask, spawn_replica_sync};
+pub use witness::{E2eeWitnessCancellation, E2eeWitnessClient, E2eeWitnessConfig};
+pub use witness_watch::{WitnessWatchTask, spawn_witness_watch};
