+++ packages/mobile-bridge-rn/tsconfig.json
@@ -3,7 +3,9 @@
   "compilerOptions": {
     "allowJs": false,
     "jsx": "react-jsx",
+    "lib": ["ES2023"],
     "noEmit": true,
+    "skipLibCheck": true,
     "types": ["react", "react-native"]
   },
   "include": ["src/**/*.ts", "src/**/*.tsx"]
