+++ plugins/db/Cargo.toml
@@ -14,6 +14,7 @@ anlg-db-core = { workspace = true }
 anlg-db-execute = { workspace = true }
 anlg-db-migrate = { workspace = true }
 anlg-db-reactive = { workspace = true, features = ["specta"] }
+anlg-db-sync = { workspace = true }
 anlg-e2ee = { workspace = true }
 anlg-fs-sync-core = { workspace = true }
 anlg-session-ingest = { workspace = true }
